+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     pages
   - Add tags to apps overview tab
   - Improve layout of apps tabs and add deployment statistics to overview tab
+  - Improve on hover pop-up behaviour for tags
 
 ## [2.20.0] - 2021-08-04